requirements.txt
boto3==1.36.2
celery==5.4.0
drf-yasg==1.21.8
Django==5.1.4
django-cors-headers==4.6.0
django-environ==0.11.2
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-elasticsearch-dsl==8.0
django-filter==24.3
elasticsearch==8.17.1
kafka-python==2.0.2
psycopg2-binary==2.9.10
Pillow==11.1.0
locust==2.33.0